+++ src/shared/auth/user-role.enum.ts
@@ -3,5 +3,5 @@ export enum UserRole {
     ADMIN = 'Admin',
     EMPLOYEE = 'Employee',
     VIP = 'VIP',
-    SUPPORT = 'SUPPORT',
+    SUPPORT = 'Support',
   }
